requirements.txt
requests>=2.25.0
pillow>=8.0.0
pyinstaller>=4.0
websockets>=10.0
aiohttp>=3.8.0
